:building_construction: [#171] Rework the data management for edit grid items
For edit grid items without isolated editing, there are no problems,
they already get their data path from the root formik context to the
leaf and there are no nested contexts.

However, in isolated edit mode, there is a nested formik context to
only commit the changes when the user clicks the save button. Initially,
we thought we can just naively render the nested components with their
names and things would be fine because Formik isolates the form
values for each item, but that breaks for radio inputs. The React/Formik
state is fine, but the browser is not aware that these things are
isolated from each other because the 'name' attribute is used to define
radio groups.

To address this, the name attribute is now made unique by prefixing it
with the (parent) editgrid key and item index, yielding a pattern
like 'name=outer:0.inner:3.myRadio', mitigating the browser behaviour.
All places that operate on the data now need additional processing to
extract the useful information from this container and apply the
right merges for conditional logic evaluation. Everything is even
harder to reason about now :(

Alternatives considered:

* rendering the 'subform' in a portal in another node, so that it could
  be wrapped in a separate <form> tag to also create the isolation.
  This is not viable, because it messes with the accessibility tree and
  would be super confusing for screenreader users.
* 'flat' prefixing - not really possible because we still need to sort
  out the root keys to update them
* modifying the formio definition keys dynamically - it's a recursive
  tree structure and there are conditional references to other keys, I
  did not feel confident that it would lead to a simpler solution