:ok_hand: [#169] Review comments
Instead of using a storybook render component with a useState, the modal close story now uses a decorator which updates the story args/controls. Because the tests controls are updated during the tests, and storybook doesn't reset its controls between tests, you have to make sure to keep the state up-to-date "manually".

The `UtrechtIcon` has some effect on our icon component when used in utrecht buttons.
Utrecht buttons use a line-height of 1.333, which nicely showcases textual content. Font-awesome icons are not affected by this, as they use a line-height of 1.

`UtrechtIcon` does not touch the line-height, creating extra empty space. This is most noticeable when the only button content is an icon.

To correct this styling issue, the modal close button uses a line-height of 1. This does not 100% fix the styling, as the UtrechtIcon still adds 1 extra pixel of spacing. But overriding the `UtrechtIcon` would remove the availability of (at least) one design token.