:wheelchair: Prevent a label tag inside a fieldset legend
Labels should be programmatically associated with input elements,
with a clear one-to-one relationship. In the case of fieldset
legends, we only want the *appearance* of a form field label,
but not the semantic meaning.

Rendering a span instead of a label tag improves accessibility
and reduces errors reported in the browser dev-tools.