:alien: Work around Formik's wonky error state
Formik tracks validation errors in the state (simplified) as key-value
pairs where the key is the name of the form field and the value an
error string. This structure is recursive - field names with a period
are nesting levels, so a value can also be another object with a
similar shape. That's also what their type definitions claim,
so far, so good.

However, Formik also has FieldArray, where the form field values
state maps to a key-value mapping of string keys (field names,
same nesting behaviour) and values of *array* type for multi-value
fields. Each browser field properly exposes . or []
access, which is fine. BUT, the errors state also has this
shape, meaning that validation errors are now key-value
mappings of strings to arrays of strings, with the error message
for the item at each index.

Except - that's not entirely true: the collection/array as a whole
may also have a validation error (e.g. max number of items, or no
duplicates allowed etc.), which maps to a string type value instead
of an array of strings, and that is something that cannot be handled
properly out of the box because accessing . will just
happily give you the character of the error message at position
, even though there is no item-level validation problem.

The workaround for this is a custom hook that extracts the error
information, aware of the single/multi-value field semantics and
it performs type introspection at runtime to check if it's an
array of error messages or a string belonging to the whole
container. For this, we use the low level getIn helper from
Formik (which it uses itself under the hood in its hooks) to
handle the '.' separator/nesting behaviour. It's conveniently
typed to return 'any', which works well in our case since we
can define/assign the actual real runtime types without also
having to fight Typescript.