:sparkles: [#43] Ensure component.conditional/clearOnHide and input multiple cooperate
While the low-level MultiField has logic to automatically insert an
empty item if none are present _on first render_, this doesn't play
well in the full FormioForm component when combined with components
that are conditionally visible with clearOnHide behaviour enabled.

The reason for this is essentially a state-update race condition in
React, specifically in the Formik values state. Roughly what happens
is:

1. FormioForm renders and evaluates the conditional logic. The end
   result is a conclusion that Formik values need to be updated to
   remove the data entirely of the hidden component.
2. Because 1. happens during render, with the Formik.values and
   component configuration at that time, the Formik.values update
   cannot be 'persisted' immediately. It is processed in a subsequent
   useEffect.
3. The FormioForm component returns JSX to render, and rendering
   continues. One of the child nodes is the MultiField component for
   the textfield with multiple: true.
4. The textfield (and thus MultiField) render, and they see the
   'current' Formik.values, because the update is scheduled for
   useEffect which runs after rendering (crucial!).
5. The MultiField detects that the value for the field is an empty
   array, and schedules an effect through useEffect to insert an
   item.
6. The MultiField JSX renders, finalizing the component render phase.
7. Now the MultiField useEffect-scheduled effects render, which leads
   to the item being inserted in the Formik.values state. Crucially,
   useEffect is called inside-out after rendering.
8. Now the FormioForm scheduled effect to update the values from the
   conditinal logic/clearOnHide is called, because all child components
   completed their life-cycle (remember: inside-out order).
9. This effect however, has a stale reference to what the values should
   be - the values update from 7. is not visible here, and effectively
   gets undone.

There's no simple way coordinate these value updates, since React's
paradigm is that state/props act as input, and there's no cleaner way
to update the Formik.values other than relying on useEffect here.

However, we can avoid this situation entirely by taking the multiple
option into account when deriving the initial values for each component,
by not needing to run the insertion effect in MultiField. And that's
precisely what this commit does :)