:alien: [open-formulieren/open-forms-sdk#863] Handle editgrid.defaultValue equal to null
The intrinsic value of an editgrid is an array of objects, but
something is going wrong in the formio-builder or formio.js layer
when an editgrid component is added, leading to defaultValue being
initialized to 'null'.

We can detect this in the initial values derivation and normalize it
in the renderer, even though it should be fixed in the builder aspect,
but that also requires patching up the backend data for existing
component definitions.