:building_construction: [#76] Rework validation schema builder interface
The interface is updated to take a context object rather than receive
each piece of context as a positional argument. The context is used
to propagate the injected dependencies.

As the actual plugin validation callback is generic in nature, a curried
function is passed in the context that is easily invoked in Zod's
`superRefine` method.

Existing components and tests are updated for the new interface.
Technically this is a breaking change, but the API is private and
we're not 1.0 yet so now is the time to make this change.