:sparkles: [#122] Using editgrid components for children display and editing
Because the children component is very similar to the editgrid: a repeating group of components, which support manually adding and inline editing, it makes sense to reuse the lower level editgrid components.

Because the children component has a fixed structure (each child has form fields for bsn, firstnames and date of birth), we can render the components directly, instead of using the dynamic form fields rendering logic.

This is a simplistic setup, which just implements the editgrid components. Some iterating will be done to remove duplicate code and create a nice structure for future components.