:sparkles: [#99] Implement scaffolding/layout of the file upload component
The file upload supports drag and drop (on non-mobile
devices), so a placeholder is set up for a drop zone
that will likely be implemented using react-dropzone.

We've opted for a logical DOM layout because that's
what screen readers use for the (tab) navigation/TOC,
but the CSS grid layout allows themes to reorder the
UI elements for the component, so that our legacy
theme can keep the structure with the description
and errors below the 'field'.

The validation errors require some special handling,
as the data type for the value of the component is
always an array, but it's not a classic MultiField
setup. So we need to check if the error is for the
field as a whole, or a list of error messages for
a file upload as a whole, or even errors for deep
attributes of a file (possibly reported by backend
validation errors).

No actual value update/upload behaviour is present
yet, this will be for a separate PR.