:sparkles: [#99] Integrate the UploadInput component into the File UI
Translate the component options into relevant props for the uploads.

Note that we must parse the fileMaxSize human readable string, as
Formio does. This was copied over almost as-is from Formio.js for the
builder *and* we have legacy configurations that must be supported.

We also pass down the input ID so that clicking the label activates the
file upload input too.

The accept prop is built from the array-format provided mime types set
in the upload component. We also have a list of file extensions, but
currently we're not doing anything with that. We disregard the
calculated filePattern prop, as it can be inferred from mime types
and extensions.