:sparkles: [#99] Translate file drop event in Formik state update
Whether the file drop was accepted or rejected, the file is always
added to the formik state with the appropriate state to display in
the file list.

We use FieldArray to manage the items inside the field state. The upload
process is async, and may also result in errors. Any observed errors
are added to the Formik error state for the particular item.

We can't rely on zod for the schema validation here, as some errors are
produced by the react-dropzone library already or may originate from
the backend. So, we must manage the error state carefully. This also
means that removing files must reshuffle the errors array and make
sure to keep things in sync.