:test_tube: [#122] Uneditable isolation editgrid item cannot retrieve value from formik context
When an editgrid is in isolation mode, only editable editgrid items can retrieve their value from the formik context.

This is caused by a combination of factors:
1. The data of the editgrid item body is fetched using a `namePrefix`.
2. When in isolation mode, the `namePrefix` is set to a special isolated value: `parentKey:itemIndex`.
3. When the editgrid item is editable, a new formik context will be created, providing access to the isolated `namePrefix`.

So, when an editgrid is isolated, but one of its items cannot be edited, the formik context won't be initiated. Causing `namePrefix` to point to nothing.