:recycle: [#99] Rework UploadInput interface
Changed callback to receive a collection of dropped files rather
than invoking it one-by-one, which enables the parent to operate
on the collection as a whole and reduce the number of concurrent/
racing state updates.

The onBlur handler is also added so that the field can be marked
as 'touched' in the Formik state for proper invalid/error display.