:safety_vest: [#99] Implement upload validation schema
Validates the max file size, maximum number of files
and the client-side upload state as part of the Formik
validation schema.

Additionally, some helpers and error messages are defined
for the eager validation feedback produced by
react-dropzone when invalid files are dragged (or selected
in the file picker UI).

The validation error messages are coordinated between the
two.

There is a situation where a file-level validation error
messages can jump to the field-level schema validation,
this happens for maximum number of file violations. it
is considered acceptable behaviour.