:safety_vest: [#178] Implementing customerProfile validation schemas
Implementing the basic validation schemas for the new customerProfile component.

It validates the existence of email and phone number data, based on component configuration set by the `digitalAddressTypes` property. The digital addresses are optional/required based on the component wide `required` validation rule.

One edge-case has been covered: using a digital address as "one time only" and as "preferred". This is a situation that should not be possible in the UI. Still, as a sanity check, the situation is included in the validation schema.