:zap: [#99] Optimize bundle chunks
Use lazy-loading mechanisms from core React with Suspense to give
bundlers the opportunity to use a separate chunk for the file upload
component and its dependencies.

If a form (step) does not use the file component type, we can defer
loading that until it's necessary, which should reduce the bundle size
of the entrypoint.