:sparkles: [#122] Children component with 'enableSelection'
Allow users to select which children should be included in the form submission.

`selected` is not a regular part of the children data. Using the checkbox component for this property, will result in weird behaviour (as there is no default value, so formik will treat it as a list of values).

The `selected` is represented using a custom checkbox component. This component sets the defaultValue to "false", ensuring that checking the input will result in a `true` value, instead of an array.