:safety_vest: [#122] Implemented validationSchema and isEmpty for children component
The children component validationSchema exsists of two parts:
- The children component validationSchema
- The children component sub-form validationSchema

Because the children component uses editgrid in isolation mode, we need to add a validation schema for each child. For the global validation, we use the "complete" validation schema.

The "sub-form" validation schema covers the validation of the individual fields: bsn, firstNames and dateOfBirth. The "complete" validation schema covers the validation of the entire component, with additional validation rules to make sure that all children data combined is valid (no duplicate bsn values).