:safety_vest: [#122] Using validationSchema in children component
Because we render a nested/isolated formik form for the editing of the children data, we have to manually set the `touched` state. So, everytime child data is changed, we mark its subfields as touched.

To detect the touched state of the Children component itself, we check if any children are marked as `touched`.

When implementing, some minor issues were noticed in the validation schema, and have been updated. These include:
- correctly setting global error messages
- using the same error messages as the SDK implementation
- marking the Children component as optional