:lipstick: Apply utrecht-document styles imperatively
Because the modal gets appended to the storybook root, it doesn't
get the design tokens from utrecht-document, resulting in font
family values not being representative and misleading snapshots.

This approach ensures that the NL DS outer container also applies
to the dialog and avoids review comments that will need
explaining over and over again.