:recycle: [#122] Splitting the children edit and create modals
By having a dedicated "edit child" and "add child" modal, we can simplify the onChange behaviour, and the overall code in the `index.tsx` file.

The ChildrenTable ChildEditButton component has been replaced with an `EditChildModal` component. This renders the edit button, and when clicked toggles an edit child modal.

The "add child" modal remains in the global code. Where we now only need to track if the modal is opened, as we know each new child should start from the `EMPTY_CHILD` data.