:sparkles: [#178] Implement generic digital address component and "add new address" feature
Because 99% of the digital address fields are (at this moment) the same, we can centralize the logic into one component. This component handles the displaying of the correct form fields with labels. (Initially I added "fieldLabel" and "addAddressButtonLabel" properties, so that the individual digital address components decided which labels were used, but later decided to keep it simple for now. If we at some point need more flexibility, we can always add properties)

Additionally, users can now decide to add a new digital address. When there are prefilled digital addresses, we show a select input. Below this input we showcase a "Add digital address" button, which changes the input to a textfield.

When in "textfield mode" the user can also change their preference. By clicking a "update your preference" button, a modal will appear with the preference options: "use only for this form" and "save for later forms".