:globe_with_meridians: [#178] Using digitalAddressType as a message variable
To simplify the multiple translatable messages for the different digital address types, all digital address type-specific messages now work with a `digitalAddressType` variable.
This allows us to maintain a limited set of messages, for the different digital address types. Instead of creating the same messages over and over, with minimal differences.