:white_check_mark: [#36] Add/update stories for marker/validation behaviour
The formio editgrid works with isolated mode, which requires us to
mark the data of an item with the expanded state so that it can be
detected in the validation schema.

The low-level component is responsible for managing this marker, while
the high-level registry component is responsible for the validation
schema to check this marker.

The types and constants are already updated for the interface exposed
to the outside world w/r to this marker.

Note that ideally we'd use a Symbol for this rather than a private
string key/property name, but that's not possible due to a limitation
in Zod: https://github.com/colinhacks/zod/issues/2734