:sparkles: [#36] Display item-level validation errors
Validation errors that belong to the item as a whole are now displayed
for the item, and appropriate item-level error styling is added to
mark just the offending item.

The error is displayed at the bottom for consistency with other
components, and we'll update the 'error under label' situation in one
pass for all components.