:safety_vest: [#36] Ensure that 'required' validation works as expected
We carefully need to inspect the error passed to the item - if there are
editgrid-level errors (a string), using the indexed access for the item
would grab the substring at that index.

This also means we don't need specific item-level errors as set up in an
earlier commit, we can just grab it from the Formik errors and do
runtime type checks to correctly interpret it.