[#221] Added disabled prop to the selectboxes
Sometimes it's needed to disable a checkbox for some reason. For example
the when a max limit of allowed checked checkboxes is reached we disable
the rest of them. The custom error message for the maxSelectedCount has
not been removed because we may have an edge case were it will be shown
and useful.