:art: [#102] Tweak validation and document microtask reason
We don't want to validate the entire form, so the setTouched should
not trigger full validation. Instead, we ensure we do validate the
single field.

The microtask approach is unclear why it's behaving the way it does
in storybook... Let's see if this becomes a problem or not, but at
least write some docs about it.