:bug: [#237] Fix display value of number component not being cleared when field value is set to null
The documentation for the `value` prop of the `NumericFormat` component (https://s-yadav.github.io/react-number-format/docs/props#value-number--string)
does not list `null` being a valid value, but the type definitions do suggest that you can pass `null` to it. In any case, passing `null` does not
clear the display value, which was already reported quite some time ago (https://github.com/s-yadav/react-number-format/issues/727).

For now, we can apply a work-around that sets the value to an empty string instead.