:recycle: [open-formulieren/open-forms-sdk#863] Refactor generic FloatingWidget into specific DatePicker widget
The interaction code is now reusable and only interacting with the
trigger icon should affect the datepicker popout state. However, we do
want to position the dialog relative to another element (the text box)
and have the option to add additional UI elements to the datepicker
for the datetime field.