:bug: [#244] Add missing aria-labelledby to inputContainer fields
Using aria-labelledby we can provide a11y context about the prefix and suffix to the inputContainer input element.

Browser give the highest precedence to `aria-labelledby`, when calculating the input label. This means that all other labels (`<label htmlFor>`, aria-label, etc.) are disregarded. To ensure that all expected labels are added to the InputContainer input, we have to include the id of the label element into the `aria-labelledby`.

For consistenty reason regarding the currect SDK implementation, and the NL-design system implementation, is the input id also included into the `aria-labelledby`.

For passing the `aria-labelledby` prop to the children element a `renderItem` setup was decided. Using React `cloneElement` we could simplify the usage a bit, but this would then "magically" assign the `aria-labelledby` prop to the input. In favor of predictability, I went the `renderItem` route.
React docs about cloneElement and its pitfalls: https://react.dev/reference/react/cloneElement