:recycle: Update reference stories to not require storyfactory
The helper was used to create identical stories for the own
implementation and the reference formio.js/react-formio implementation.

We're removing the latter and keeping the former, as the reference
tests have done their job to guard feature parity, but now we want to
drop the formio.js dependency and possibly (deliberately) diverge in
behaviour.