:bug: [#5829] Correct profile subfield usage
The profile component subfields should be used based on the pre-populated data, and the current component value.

When the `useDigitalAddresses` hook is used, and the component addresses are empty, the addresses are updated to one of the fetched addresses. Then, in the UI, we only show the dropdown if there are pre-populated addresses, and the current component value is of a pre-populated address.

This scenario can easily be explained with moving between multiple form-steps:
If you go to a previous form-step, you expect to see the values you entered. When you used the dropdown, you want to see the dropdown again. But, if you used the textfield, you would expect the textfield to be present again.

Example of using textfield:
1. You open a form. Addresses are fetched, the first address is selected as default, and the dropdown is shown. You change the UI to textfield and enter a completely new address.
2. You go to the next step, and re-think your entered address. You go back to the first step.
3. Addresses are again fetched. Because you already entered a value, the first fetched address isn't set as default. And, because your manually entered address isn't part of the fetched addresses list, we show the textfield.

Example of using dropdown:
1. You open a form. Addresses are fetched, the first address is selected as default, and the dropdown is shown. You change the value to the second fetched address.
2. You go to the next step, and re-think your entered address. You go back to the first step.
3. Addresses are again fetched. Because you already entered a value, the first fetched address isn't set as default. Now, because your selected address is part of the fetched addresses list, we show the dropdown.