:safety_vest: [#5824] Simplify profile component `showPreferencesButton` logic
Instead of calculating the validity of the parent component, we can simply pass it along from the parent to the child.

The touched state in the `showPreferencesButton` logic doesn't serve any purpose. It was intended as "only show the button after human input", but this will (at this time) always be the case with the textfield subfield.

The touched state could even cause issues/confusion with moving between form-steps. Where the user goes to a previously filled-in step to adjust their profile preferences, but the "update preferences" is for some reason hidden. Only after focus/blur will the button be shown.

The logic should be: "does the field have a valid value for the intended digital address type?"