:test_tube: [#274] Failing tests for early datepicker validation trigger
The datepicker validation shouldn't kick-in when the focus is still within the component. So placing focus on the datepicker input, and pressing tab to move focus to the datepicker trigger button, should not trigger validation. Only after the focus moves outside the bounds of the datepicker component, should validation trigger.