:bug: [#5817] Fix initial value of digitalAddressPreferencesModal
The initialValue of the `digitalAddressPreferencesModal` didn't take the actual filled-in preference into account. This resulted in issues when using the component in a multi-step form.

When entering a new address, setting the preference to "use as preferred", going to the next step, going back to the profile step, and opening the preference modal again, the preference would be reset to the default value (use only once). The actual value of the preference wasn't affected, so when submitting the form the preference was set to "is new preferred". This is/was purly a visual bug in the preference modal.